requirements.txt
pillow
pyautogui
colorama
